pretty_bad_protocol: Support exporting encrypted secret keys
As part of the Sequoia migration, we want to export secret keys out
of the GPG keyring and into our new database storage.

So, have the gpg.export_keys() function accept a passphrase to export
encrypted keys that GPG wants a passphrase for. Internally we switch it
to use the `_handle_io` function since that has support for taking
passphrases; the one weird thing is that it requires an input file (e.g.
if you were encrypting a file) but since there's no input here an empty
temporary file works just fine.

For whatever reason I couldn't get a simple `p.stdin.write(passphrase)`
to work, it would hang on stdout, which is probably related to the
pre-existing comment about "stdout will be empty in case of failure"...

A new test generates a new GPG key pair, exports the public and secret
keys from the keyring and then encrypts and decrypts a message using
Sequoia to test compatibility. It also checks a few error cases
like invalid fingerprint and missing secret key.

Refs #6802.