chore: factor out minimal, specialized template for changes to production dependencies
We're good at filling in the diff review for meaningful updates to
production dependencies, less good at leaving good breadcrumbs for why
we choose to silence certain alerts for them.  Factoring out a
specialized template for the latter is an experiment to try prompting
specifically for the relevant information, with the bonus of simplifying
the template for everything else.