Delay apache2 restart in postinst
If we restart Apache before database updates are applied, then we have
the possibility of running into missing columns or tables for a short
time. Also there's a theoretical possibility that users could try to
submit data mid-migration, so let's just have restarting apache be the
very last thing we do.

Co-authored-by: Cory Francis Myers <cory@freedom.press>