Add script to validate exported journalist pubkey
Add a new script that is similar to admin/bin/validate-gpg-key.sh,
except this also asserts that there's only one PGP key in the file.
This will be used in a follow-up to verify the key file we create
matches what we're looking for.

Ideally this should use Sequoia instead of pretty_bad_protocol, but that
takes some more work and probably makes sense to do in a separate PR.