Fix demo build
Build the redwood wheel during the container build instead of when
launching the container. Copy some of the rustup logic into the
DemoDockerfile and use the same base image version as the current dev
environment.

This is hopefully good enough for now, long-term we should switch to
server nightlies that are just installed into the image.