Migrate public keys from GPG to database-backed storage
Add an alembic migration that iterates over the GPG keyring, identifies
source keys, exports them from GPG and saves them into the database.

The main failure risks are the interactions with GPG. We already run
`gpg.list_keys()` on startup, so it's unlikely that's broken (and if it
is, we have bigger problems). So the main concern is exporting the key
might fail. The export operation is wrapped in a try/except and we
validate the exported key we get from GPG.

Notably, this does not increase our footprint of pretty_bad_protcol
usage as the two functions being used are already in use elsewhere in
SecureDrop.

Some higher-level design information is at
<https://github.com/freedomofpress/securedrop/pull/6946#discussion_r1343339398>.

Fixes #6800.