redwood: Add is_valid_public_key function
We have multiple cases in Python code where we want to validate that a
string is valid PGP public key that Sequoia will accept. Exporting a
`is_valid_public_key` function will help make checks more robust and
simplify some tests.

The validation code on the Rust side is also a bit duplicated and
currently manually kept in sync with comments, so let's take the
opportunity to unify it into one file and de-duplicated with a macro
(hopefully not too magical).

Tests are provided that verify a good public key passes, a bad SHA-1
key is rejected and a secret key is also rejected.

Fixes #6947.