Migrate secret keys from GPG to database-backed storage
For the final part of the GPG->Sequoia migration, we export sources'
secret keys out of GPG and into database-backed storage for use with
Sequoia/redwood functions. Because we don't have access to a source's
GPG passphrase, we can only do this migration when they log in.

If everything has been successfully migrated, then we delete the GPG key
out of the keyring as well.

Tests have been added that cover the pre-existing keypair generation
code and the new secret key migration.

Fixes #6802.