Fix other migrations to just catch SecureDropConfig usage
Wrapping the entire migration in a try/except is incredibly broad, given
the risk of silently suppressing errors in production and marching
forward regardless. In these cases, the migrations are avoiding running
for fresh installs, where config.py isn't setup.

We can detect that case explicitly and allow the rest of the code to run
normally.