Test alembic migrations when config.py doesn't exist
alembic migrations run in two contexts, 1) upgrade, when config.py
exists, and 2) fresh install, with no config.py. We only test the former
scenario, so this adds tests for the latter. Each migration is run up
and down with a config.py and without a config.py, to catch any
invocations of SecureDropConfig that aren't properly guarded.

This would've caught a regression like #6976.

Fixes #6977.