redwood: Improve writing of encrypted files
First, use "create_new" mode when opening the file so if it already
exists, an error will be raised instead of silently truncating the
existing file. This required adjusting our tests a bit since
NamedTempFile creates a file before we try writing to it.

Then use a BufWriter to buffer filesystem writes instead of writing each
byte as it comes in.

Fixes #6989.
Fixes #6990.