redwood: Don't armor encrypted files/messages
...except in one test.

We didn't have GPG armor ciphertext, so we don't need redwood to either.
This will result in a noticable speedup during encryption. Keys will
still be armored when being stored in the database.

Tests make this a bit more complicated, we want to generate an armored
message we can post to the journalist API, so we make it an optional
parameter for redwood.encrypt_message() only. We have to adjust other
tests now that we can no longer easily look for the
"-----BEGIN PGP MESSAGE-----" string.

Refs #7022.