WIP: Migrate source public key to database storage online too
If the offline alembic migration fails or is skipped for whatever
reason, try to migrate the public key online too, right at the same time
we migrate the secret key.

A bit of refactoring into the EncryptionManager class helps simplify the
try/except necessary to keep these migrations from being fallible. This
also makes it straightforward to also try the public key migration from
the JI if we wanted.