Lower cargo vet review criteria to "safe-to-run"
For many crates we don't have the technical expertise to sign off on
"safe-to-deploy", as defined by the cargo-vet documentation. The
"safe-to-run" criteria is roughly the same standard we hold for Python
diff reviews, with the benefit that our reviews are legible to the rest
of the Cargo Vet community.

Refs #6999.

Co-authored-by: Cory Francis Myers <cory@freedom.press>
(cherry picked from commit 0d285b789b4e139b8d31c6e561f1fcfc3237d5d4)