Fix Submission Key validation at startup, improve SI experience
The previously implemented Submission Key validation during JI startup
did not work in production because it was under the
`if __name__ == "__main__":` block, which isn't executed by WSGI.

In the JI, we now run it upon import and exit if it's not valid.
For the SI, we check during initialization, and set a config flag.
Each request checks said flag, and if invalid, render a new
`offline.html` template as a HTTP 503.

Testinfra smoke tests cover behavior on prod instances, but since it's
destructive, will only run in staging.

Fixes #7058.

Co-authored-by: Kevin O'Gorman <kog@freedom.press>