build: Fix cargo-vet binary caching
The upstream cargo-vet CI manifest was a bit overcomplicated,
trying to install to a different path and it just wasn't working.

Instead, install cargo-vet to the normal path, store it in the
actions/cache, and skip installing on cache hits.

On a cache hit, it drops job runtime to less than a minute.

Fixes #7011.