Fixed several issues related to the dependency check in bootstrap.py
There were a few areas of broken logic in the is_missing_dependency
function, which would result in broken or unidentified behavior if
the underlying apt-cache was stale or non-existant.

If, while doing a search of the apt-cache, it only located a
subset of the package list, or if it were completely unable to find
a package (such as in the event of a missing cache), it would still
return False, because the grep operation didn't find the "Installed:
(none)" string it was looking for. So in those situations, even
though the dependencies were missing, it did not return True as it
should have.

The way the dependencies were listed at the top, with spaces
inserted instead of tabs, meant that when we split the string using
a space character as the delimiter to form the command, there were
dozens of empty strings, which apt-cache would attempt to search for,
and would turn up the same error as a missing package,
making it impossible to determine when packages were missing.

This change eliminates those extra spaces, simplifies the logic
to prefer built-in string-checking in Python rather than shelling
out to grep, and fixes the broken return logic so it identifies
when dependencies are missing due to not being in the apt cache
at all.