Persist onion addresses with "make dev-tor"
Makes the SecureDrop Workstation dev workflow easier by removing
the need to setup standalone server VMs if one wants a persistent
server address and authentication token.

The solution involves creating a docker volume. This will persist
as long as the user does not manually remove it or moves to another
system without porting the volume. Since this a dev environment,
long term persistence should not be a concern, so this solution
sufices.

Fixes #7115