DNM: Power `make dev-tor` with Arti
The authenticated onion service stuff is still unmerged
(<https://gitlab.torproject.org/tpo/core/arti/-/merge_requests/2266>),
so we build Arti from the MR and it works! Generating a config.toml file
seems much cleaner than adding to `torrc` and generating `.auth` files.

The main upside for this is that using Arti is cool and gets it more
real-world testing. But the downside is that it takes a non-trivial
amount of time to build Arti locally and maybe it's not worth doing
until we have pre-package/downloadable binaries. :thinking: