Validate user provided same password back to server
We require that all users have a strong diceware password by generating
it ourselves for them to use. However, we accept any diceware-looking
sent in the POST request, meaning that a user can deliberately set a
weak or fixed password. (It would be difficult for an attacker to do so
since they'd need to both bypass the onion service authentication and
CSRF.)

To fix this, upon generating a new password for users, we save a hash of
it in the user's session. When they save the password, we validate it
against the hash and if matches, allow it to be saved. This ends up
being mostly straightforward (outside of tests), requiring modifications
to utils.set_diceware_password() and the creation of
utils.set_pending_password() and utils.verify_pending_password().

Tests are a bit more complicated because we rely on being able to send
arbitrary POST strings to the correct endpoint, so it needs to manually
be set in the session. New tests for this functionality are added as
well.

This addresses "SEC-01-002 WP4: Possible Account Takeover via Weak
Password Policy (Low)"

Fixes <https://github.com/freedomofpress/securedrop-security/issues/94>.