Use upstream gcloud-sdk image instead of our oudated one
We have not updated our gcloud-sdk image since 2018 and it's really not
worth us maintaining this when Google provides its own image.

We now authenticate using the method documented at
<https://cloud.google.com/sdk/docs/downloads-docker>, in which we invoke
the container to login, which creates a volume that can be used in
future calls. This means we no longer need to have a running container
that we exec into; we can just create a new container each time we want
to make a gcloud call.

Fixes #7195.
Fixes #7196.