Migrate staging job to GitHub Actions; remove CircleCI
We are officially at the end of the road for the CircleCI to GitHub
Actions migration! This migrates the final job over and removes the
manifest entirely.

The staging credentials have already been configured by infra, all that
was needed was adjusting some of the `$CIRCLE_` environment variables. I
added a bit of output to aid future debugging.

Fixes #6886.