Migrate static-analysis-and-no-known-cves to GitHub Actions
This ports the same functionality but in a much simpler way since we
just need the latest versions of the safety and semgrep packages.

We want this to run on each PR and every night, so move it into
security.yml and configure that workflow to run on each PR as well. As a
side-effect, `rust-audit` will now run on each PR, which is fine.