Migrate translation-tests to GitHub Actions
Translation tests are slow because it's our normal tests just run 20+
times, once for each language. We can use a dynamic matrix to generate
separate jobs pretty easily (I learned this pattern from
https://github.com/suzuki-shunsuke/example-github-actions-dynamic-matrix).

This job is a little special as we only run it in specific
circumstances. Two are easy to handle 1) branch filter and 2) scheduled
run.

But only triggering on PRs from "weblate-fpf" is a bit harder, so we add
some run-time logic in the locales job itself.