Upgrade ruff, remove black
Upgrade ruff to the latest version so we can use it for formatting and
native editor integration. While we're at it, introduce the standard
`make fix` target
(<https://github.com/freedomofpress/securedrop-tooling/issues/11>).

We can remove `--show-source` too now that it's the default.

Actually fixing new issues will happen in follow-ups to make this change
easier to review.

Fixes #7226.