Speed up `update-python3-dependencies`
The `make update-python3-dependencies` step is slow for two main
reasons: container image layering and pip-compile.

Switch to the new uv tool, which reimplements pip-compile in a much
faster way. The output is basically the same, except the sorting is
smarter (e.g. pytest comes before pytest-cov) and package names are
properly normalized. We can also drop the `--allow-unsafe` because uv is
entirely independent of setuptools and pip-tools.

Because uv is independent of the system Python (it's pure Rust), we can
run it on the host and outside the container, so no image rebuilding
happens during the updates, it'll just be needed once afterwards.

uv is still quite new to the Python ecosystem, but this allows us to
begin using it without any lock-in, it should be trivial to swap back to
pip-tools if needed.

Overall `make update-python3-dependencies` now takes seconds to run
instead of minutes \o/

Refs <https://github.com/freedomofpress/securedrop-tooling/issues/16>.