Use a single pass in ansible to install local packages
In edbc81539582, the process was refactored to use a two-pass system in
which packages were first installed with `apt`, which would resolve
dependencies, and then `dpkg`, which forcibly installed/overwrote
existing packages.

Recently we've been hitting trouble where the dpkg lock is still held
and the second pass fails.

We can do this all in one pass using apt though, by installing all the
packages at once, so the dependency graph can be fullfilled, and by
passing `--reinstall` to force the installation of the new debs.

Fixes #7258.