Fix conffiles hack to not generate invalid listings
`echo ""` adds a single blank line, which dh_conffiles treats as an
invalid file, which is an error in noble. If we have a fully empty file,
it is happy and still applies our hack.

Remove the duplicated dh_gencontrol stanza that was already being
overridden later while we're at it.