Copy backup to Admin Workstation before manual noble migration
If the admin has manually initiated a backup, let's copy a backup to the
Admin Workstation so it's easier to use. The backup is automatically
deleted at the end of a successful playbook run.

Especially for instances using SSH-over-Tor, this step may take a while,
but if something does go wrong, it'll be nice.

A backup is still taken on the server that stays on the server, there's
no adverse effect except for a slight delay in waiting for the
extraneous backup.